pinctrl: intel: Avoid potential glitches if pin is in GPIO mode
When consumer requests pin, we, in order to be on the safest side,
switch it first to GPIO mode followed by immediate transition to
the input state. Due to posted writes it's luckily to be a single
I/O transaction.

However, if firmware or boot loader already configures pin to
the GPIO mode, user expects no glitches for the requested pin.
We may check if the pin is pre-configured and leave it as is
till the actual consumer toggles its state to avoid glitches.

Fixes: 7981c0015af2 ("Add Intel Sunrisepoint pin controller and GPIO support")
Depends-on: f5a26acf0162 ("Initialize GPIO properly when used through irqchip")
Reported-by: Oliver Barta <oliver.barta@aptiv.com>
Reported-by: Malin Jonsson <malin.jonsson@ericsson.com>
Signed-off-by: Andy Shevchenko <andriy.shevchenko@linux.intel.com>