ISI Mailbox driver for interacting with Intel Safety Island
Intel Safety Island is a safety IP targeted for Industrial application.
The IP is responsible for following platform FuSa activities:
        a. Error monitoring of various errors in PCH and Compute Core.
        b. Functional safety test orchestration
        c. On Demand Cross Comparison(ODCC)
        d. Fault robust WCET and BCET timer services

This driver is primarily meant for interaction with ISI FW for ODCC and
for getting async notification from ISI FW in case ISI detects a fault
and on WCET/BCET timer expiry etc.

To interact with ISI, the User applications are expected to register
with ISI FW as Workload.
ISI FW assigns a workload_id to each registered workload.
By design, any user application can register as a Workload.
Each child that thread the registered workload spawns share the same workload_id.
The driver facilitates sending command from a workload to ISI FW and
receiving and forwarding the response to the concerned workload.
The driver supports following IOCTs:
        a. Workload register: To register an app as Workload with ISI
        b. Workload deregister: To deregister a workload.
        c. Async command register: Register to get notification for async commands
        d. Async command deregister: Deregister for a async command notification
        e. Async command get: Get notification for registered async command
        f. Sync Txn Send receive: Send synchronous txn and receive response from ISI