drm/i915/gvt: Allow Sos to access the plane owned by Guest before guest is booted
Currently when one plane is assigned to UOS, it will block all the access to plane
register in SOS. In such case it will cause that nothing is displayed on some display
in course of sos booting. This is not expected.
Before it is assigned to guest, it will still allow the SOS to access the
plane register.

Tracked-On:projectacrn/acrn-hypervisor#4572
Reviewed-by: Liu Xinyun <xinyun.liu@intel.com>
Signed-off-by: Zhao Yakui <yakui.zhao@intel.com>