vhm: use 2MB compound pages for ACRN mempool
The hypercall HC_GET_PLATFORM_INFO may require a large memory buffer for
its VM config if a large number of VMs should be reported (roughly 2400
bytes are needed per VM).

Statically use the size of 2MB for each buffer, as this is enough for
any reasonable amount of VMs the hypervisor may have and also aligns
with the design of 2MB huge pages.

Use the __GFP_COMP flag to indicate the use of compound pages. Although
this may seem to be a bit of an overkill at the moment, it provides
better semantics for higher-order pages should we ever consider mapping
them to user space.

Tracked-On: projectacrn/acrn-hypervisor#4753
Signed-off-by: Peter Fang <peter.fang@intel.com>