dt-bindings: media: kmb-imx334-sensor: Add Keem Bay IMX334 sensor
The Keem Bay imx334 sensor is a Sony CMOS active pixel digital image
sensor with an active array size of 3864H x 2202V. It is programmable
through I2C interface. The I2C address is fixed to 0x1a as per sensor
data sheet. Image  data is sent through MIPI CSI-2, which is configured
as 4 lanes at 1188 Mbps.

The sensor can also be probed as platform device and later on registered
to the I2C bus. Probing as platform device is used when different
sensors with the same slave address must be able to be attached to one
slot. They are probed with priority. Example slot with multiple sensors
attached.

Signed-off-by: Daniele Alessandrelli <daniele.alessandrelli@intel.com>