virt: acrn: support asyncio
Asyncio is a new mechanism in ACRN, which is special for these devices
which need high I/O performance. ACRN hypervisor would process the I/O
request from User VM in an async mode.

For this kind of I/O reqeusts, a new shared page is created by device
model in userspace, and register to kernel with ioctl. ACRN hsm driver
will pin the page to avoid page swapping. Use hypercall to pass the base
addresses to hypervisor, so that hypervisor can put asyncio requests to
this buffer.

When User VM traps to hypervisor due to an IO/MMIO access, hypervisor
need to distinguish it between original syncronize IO and asyncronize IO.
So, the hypercalls are added to assign or deassign the aynscio info to
hypervisor.

In ACRN, we use the ioeventfd in vhost device. We also support the
virtio device using ioveventfd for FE to NOTIFY BE. But in current
implementation, when the User VM write the NOTIFY register to notify BE,
it will trap to hypervisor. After the write access is finished, User VM
can return.
To return earlier, we provide an asyncio mode for an I/O request. For
asyncio, hypervisor directly put the request in an asyncio shared buffer,
and then return to User VM. To take use of this mechanism, add a new flag
ACRN_IOEVENTFD_FLAG_ASYNCIO to indicate that the IO is asyncio. And use
hypercall to register the asyncio info to hypervisor.

When the User VM do an IO/MMIO access and traps to hypervisor, hypervisor
will parse the I/O base address, and check if it was registered. If yes,
hypervisor will queue the corresponding ioeventfd to the aysncio buffer,
and inject an interrupt to Service VM. ACRN HSM driver in Service VM
will traverse the fd list and signal the eventfd.

Signed-off-by: Conghui <conghui.chen@intel.com>
Reviewed-by: Fei Li <fei1.li@intel.com>